rename to bazaarspreadfliplist
